refactor(etl): move utility to generic package
