+++ data/processing/projects.py
@@ -2,6 +2,7 @@ from typing import Optional
 import pandas as pd
 
 from data.constants.raw_data_constants import ProjectsDatasetCols
+from data.processing.data_parser import brazilian_date
 
 
 def projects_dataset(source: pd.DataFrame) -> pd.DataFrame:
@@ -57,30 +58,6 @@ def projects_dataset(source: pd.DataFrame) -> pd.DataFrame:
               .drop_duplicates(["ID Projeto"]))
 
     return result
-
-
-def brazilian_date(date_str: Optional[str]) -> Optional[str]:
-    """
-    Converts a date string to the Brazilian date format (dd/mm/yyyy).
-
-    Args:
-        date_str (Optional[str]): The input date string.
-
-    Returns:
-        Optional[str]: The date in Brazilian format, or None if conversion fails.
-    """
-    if (date_str is None
-        or str(date_str).strip() == ""
-            or str(date_str).lower() == "nan"):
-        return None
-
-    parsed_date: pd.Timestamp = pd.to_datetime(date_str,
-                                               errors="coerce")
-
-    if parsed_date is None or pd.isna(parsed_date):
-        return None
-
-    return parsed_date.strftime("%d/%m/%Y")
 
 
 def project_status(start_date: Optional[str],
